requirements.txt
# NuniPdf 프로젝트 의존성
# 기존 PyQt6 및 관련 라이브러리
PyQt6>=6.4.0
PyQt6-Qt6>=6.4.0

# 문서 검색 기능 관련 라이브러리
PyMuPDF>=1.23.0
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=2.0.0
pyxlsb>=1.0.10
pyhwpx>=0.0.20
olefile>=0.46
